@@ -1,19 +1,19 @@
 plugins {
     id 'java'
-    id 'org.springframework.boot' version '3.5.5'
+    id 'org.springframework.boot' version '3.5.8'
     id 'io.spring.dependency-management' version '1.1.7'
     // Analysis
-    id 'org.sonarqube' version '6.3.1.5724'
+    id 'org.sonarqube' version '7.2.1.6560'
     id 'jacoco'
     id 'net.ltgt.errorprone' version '4.3.0'
     id 'com.diffplug.spotless' version '7.2.1'
 }
 
 group = 'no.hvl.tk'
-version = '2.0.0'
+version = '2.0.1'
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_21
+    sourceCompatibility = JavaVersion.VERSION_23
 }
 
 repositories {
@@ -26,17 +26,17 @@ dependencies {
 
     // XML-marshalling not part of jdk 11 anymore.
     // https://mvnrepository.com/artifact/jakarta.xml.bind/jakarta.xml.bind-api
-    implementation group: 'jakarta.xml.bind', name: 'jakarta.xml.bind-api', version: '4.0.2'
+    implementation group: 'jakarta.xml.bind', name: 'jakarta.xml.bind-api', version: '4.0.4'
     // https://mvnrepository.com/artifact/org.glassfish.jaxb/jaxb-runtime
-    implementation('org.glassfish.jaxb:jaxb-runtime:4.0.5')
+    implementation('org.glassfish.jaxb:jaxb-runtime:4.0.6')
     // https://mvnrepository.com/artifact/jakarta.activation/jakarta.activation-api
-    implementation group: 'jakarta.activation', name: 'jakarta.activation-api', version: '2.1.3'
+    implementation group: 'jakarta.activation', name: 'jakarta.activation-api', version: '2.1.4'
 
     // Add pair class
     // https://mvnrepository.com/artifact/org.apache.commons/commons-lang3
-    implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.18.0'
+    implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.20.0'
     // https://mvnrepository.com/artifact/commons-io/commons-io
-    implementation 'commons-io:commons-io:2.20.0'
+    implementation 'commons-io:commons-io:2.21.0'
 
     // Spring
     implementation 'org.springframework.boot:spring-boot-starter-web'
@@ -53,7 +53,7 @@ dependencies {
     // https://mvnrepository.com/artifact/commons-io/commons-io
     testImplementation 'commons-io:commons-io:2.20.0'
 
-    errorprone('com.google.errorprone:error_prone_core:2.16')
+    errorprone('com.google.errorprone:error_prone_core:2.45.0')
 }
 
 tasks.named('test') {
