@@ -67,10 +67,9 @@ public class BPMNEventSubprocessRuleGeneratorImpl implements BPMNEventSubprocess
         .forEach(
             startEvent -> {
               switch (startEvent.getType()) {
-                case NONE:
-                  throw new BPMNRuntimeException(
-                      "None start events in event subprocesses are useless!");
-                case MESSAGE:
+                case NONE -> throw new BPMNRuntimeException(
+                    "None start events in event subprocesses are useless!");
+                case MESSAGE -> {
                   if (startEvent.isInterrupt()) {
                     createStartInterruptingEvenSubprocessFromMessageRules(
                         process, eventSubprocess, collaboration, ruleBuilder, startEvent);
@@ -78,12 +77,11 @@ public class BPMNEventSubprocessRuleGeneratorImpl implements BPMNEventSubprocess
                     createStartNonInterruptingEvenSubprocessFromMessageRules(
                         process, eventSubprocess, collaboration, ruleBuilder, startEvent);
                   }
-                  break;
-                case SIGNAL, ERROR, ESCALATION:
-                  break;
-                default:
-                  throw new BPMNRuntimeException(
-                      "Unexpected start event type encountered: " + startEvent.getType());
+                }
+                case SIGNAL, ERROR, ESCALATION -> {
+                }
+                default -> throw new BPMNRuntimeException(
+                    "Unexpected start event type encountered: " + startEvent.getType());
               }
             });
   }
